7/15/2016 - CunningView bug fixed, CharacterSaveList implemented

+++ Detective/src/byui/cit260/detective/view/TempIntelView.java
@@ -17,7 +17,7 @@ public class TempIntelView extends View implements Serializable {
 
     private String playerIntelSkill;
     private long intelSkillValue;
-    int observation;
+    long observation;
     
     public TempIntelView() {
         super("Please enter intel skill.");
